Deleted .DS_Store file
